Implement basic "yes" command
This command will accept the filename of an application and information
about the next stage, then write that new stage to the application.

This uses `tu` to parse date/time for the deadline of the stage.

Still need to add the ability to name the stage and specify the time of
the stage explicitly.