Fix Frontmatter parsing
My previous change to short circuit after the end of the frontmatter was
completely wrong and erased content.

Opted for a quick and dirty incrementing counter which goes up whenever
we see a "---", and only reads frontmatter at count = 1. Should probably
trim this or make it more sophisticated but it works well for now.